TurtleGUI now correctly creates an ArenaPanel with relevant TurtleInfo and even draws the Arena :)
Was able to use some of the code I had from ArenaDraw back when it was in the Arena package, but since a lot has changed I wasn't able to use all of it. I've got everything drawn on one level right now, but I left the remnants of the other classes in as well as commented out the code that referred to them in ArenaDraw.

Default Line color I set to red/width 2, just so I could see better what was happening.

But-- check it out, it's pretty neat :)