Small changes to correctly update the image to reflect changes in turtle's heading.
TurtleList now supports a "maxWidth" and "maxHeight" for the image included. So if a user tries to give us a huge turtle image, it won't break the list.

Changed the default turtle model to capital letter "A" so the user can clearly see what the heading of their turtle is.

Turtles>"Add Turtle" still working, can access other turtles with "tell [ turtleID ]" changing that turtleID to your activeTurtle.