Now working. The "errors" introduced are because I had to delete view, then reload it in to stop SVN from hating me.
If you find any errors, check first that the imports are linked correctly, because they won't be. But it is about the same as the one I have been emailing as a .zip. Added the prints that I had in before, couple of methods in Arena/Turtle, and fixed references in view that weren't meshing with this version.

Images are now in, and if you run, go to "Add Turtle" to get the cool little popup menu.